+++ History.md
@@ -29,7 +29,9 @@ switch
 - validate as library doc
 - unify loggers
 - release notes
-- new tradeparty class, switch recipient, shipping address from contact to tradeparty
+- new tradeparty class, Contact getOwnContact superseded by TradeParty getSender
+- switch recipient, shipping address from contact to tradeparty
+- new invoicecorrection class
 
 ### 2.0 still todo
 - verapdf as prevalidation?